Split cpp and utils library to granulate dependencies
dbg library - contains assert and depends on a console
utils - generic utilities without external dependencies

dependent modules changed as well
