+++ lib/CMakeLists.txt
@@ -5,6 +5,7 @@ add_subdirectory(emc)
 add_subdirectory(allocators)
 add_subdirectory(fs)
 add_subdirectory(utils)
+add_subdirectory(debug)
 add_subdirectory(types)
 add_subdirectory(thread)
 add_subdirectory(containers)
