+++ lib/allocators/CMakeLists.txt
@@ -1,7 +1,7 @@
 add_library(allocators STATIC alloc.cpp)
 target_include_directories(allocators PUBLIC export)
 # Requires assert
-target_link_libraries(allocators utils)
+target_link_libraries(allocators dbg)
 # Static analysis
 add_cppcheck(allocators UNUSED_FUNCTIONS STYLE POSSIBLE_ERROR FORCE)
 # Unit tests
@@ -12,4 +12,4 @@ add_unit_host_test(NAME pool_allocator
 	tests/pool_main.cpp
 	alloc.cpp
 	INC_DIRS export
-	DEPENDS utils libcpp)
+	DEPENDS libcpp dbg)
