+++ lib/containers/CMakeLists.txt
@@ -1,5 +1,6 @@
 add_library(containers INTERFACE)
-target_include_directories(containers INTERFACE export)
+target_include_directories(containers INTERFACE export utils)
 add_unit_host_test(NAME list
 	SOURCES tests/list_unit.cpp
-	INC_DIRS export)
+	INC_DIRS export
+	DEPENDS utils)
