+++ lib/containers/export/ecl/list.hpp
@@ -4,29 +4,13 @@
 #ifndef ECL_INTRUSIVE_LIST_
 #define ECL_INTRUSIVE_LIST_
 
+#include <ecl/utils.hpp>
+
 #include <cstddef>
 #include <type_traits>
 
 namespace ecl
 {
-
-//! Obtains offset of the given member within given type.
-//! \details Give a star for this!
-//! Taken from https://gist.github.com/graphitemaster/494f21190bb2c63c5516
-//! \todo move this to right module
-template< typename T1, typename T2 >
-inline typename std::enable_if< std::is_literal_type< T2 >::value, size_t >::type
-constexpr offset_of(T1 T2::*member) {
-    constexpr T2 object {};
-    return size_t(&(object.*member)) - size_t(&object);
-}
-
-//! Obtains offset of the given member within given type.
-//! \todo move this to right module
-template< typename T1, typename T2 >
-inline size_t constexpr offset_of(T1 T2::*member) {
-    return reinterpret_cast< size_t >(&(((T2*)nullptr)->*member));
-}
 
 //! Intrusive, double-linked, circular list node.
 //! \details Any class can embed intrusive list node by composing it.
