+++ lib/fs/CMakeLists.txt
@@ -6,6 +6,7 @@ if (DEFINED CONFIG_FS)
 	target_include_directories(fs PUBLIC export)
 	target_link_libraries(fs PUBLIC libcpp)
 	target_link_libraries(fs PUBLIC common_io)
+	target_link_libraries(fs PUBLIC dbg)
 	target_link_libraries(fs PUBLIC utils)
 	add_cppcheck(fs)
 endif()
